+++ lib/service/Appointments.dart
@@ -14,8 +14,10 @@
  * limitations under the License.
  */
 
+import 'package:flutter/material.dart';
 import 'package:illinois/model/wellness/Appointment.dart';
 import 'package:rokwire_plugin/model/explore.dart';
+import 'package:rokwire_plugin/service/app_livecycle.dart';
 import 'package:rokwire_plugin/service/deep_link.dart';
 import 'package:rokwire_plugin/service/log.dart';
 import 'package:rokwire_plugin/service/network.dart';
@@ -28,6 +30,14 @@ import 'package:illinois/service/Auth2.dart';
 
 class Appointments with Service implements ExploreJsonHandler, NotificationsListener {
   static const String notifyAppointmentDetail = "edu.illinois.rokwire.appointments.detail";
+  static const String notifyAppointmentsAccountUpdated = "edu.illinois.rokwire.appointments.account.updated";
+  
+  static const String _appointmentRemindersNotificationsEnabledKey = 'edu.illinois.rokwire.settings.inbox.notification.appointments.reminders.notifications.enabled';
+  
+  DateTime? _pausedDateTime;
+  
+  AppointmentsAccount? _account;
+  bool? _isLastAccountResponseSuccessful;
 
   List<Map<String, dynamic>>? _appointmentDetailsCache;
 
@@ -41,6 +51,8 @@ class Appointments with Service implements ExploreJsonHandler, NotificationsList
   void createService() {
     NotificationService().subscribe(this, [
       DeepLink.notifyUri,
+      AppLivecycle.notifyStateChanged,
+      Auth2.notifyLoginChanged
     ]);
     Explore.addJsonHandler(this);
     _appointmentDetailsCache = <Map<String, dynamic>>[];
@@ -55,8 +67,31 @@ class Appointments with Service implements ExploreJsonHandler, NotificationsList
   }
 
   @override
+  Future<void> initService() async {
+    await _initAccount();
+    await super.initService();
+  }
+
+  @override
   void initServiceUI() {
     _processCachedAppointmentDetails();
+  }
+
+  // Getters & Setters
+
+  AppointmentsAccount? get account => _account;
+
+  bool get isAccountValid => (account != null) && (_isLastAccountResponseSuccessful == true);
+
+  bool get reminderNotificationsEnabled =>
+      isAccountValid &&
+      (Auth2().isLoggedIn) &&
+      (Auth2().prefs?.getBoolSetting(_appointmentRemindersNotificationsEnabledKey, defaultValue: true) ?? true);
+
+  set reminderNotificationsEnabled(bool value) {
+    if (Auth2().isLoggedIn) {
+      Auth2().prefs?.applySetting(_appointmentRemindersNotificationsEnabledKey, value);
+    }
   }
 
   // APIs
@@ -132,6 +167,95 @@ class Appointments with Service implements ExploreJsonHandler, NotificationsList
     }
   }
 
+  Future<void> _initAccount() async {
+    await _loadAccount();
+    if ((account == null) && Auth2().isLoggedIn && (_isLastAccountResponseSuccessful == true)) {
+      // if the backend returns successful response without account, then populate the account with default values
+      _initAccountOnFirstSignIn();
+    }
+  }
+
+  Future<void> _loadAccount() async {
+    if (Auth2().isLoggedIn) {
+      String? url = "${Config().appointmentsUrl}/services/account";
+      http.Response? response = await Network().get(url, auth: Auth2());
+      int? responseCode = response?.statusCode;
+      String? responseString = response?.body;
+      if (responseCode == 200) {
+        _isLastAccountResponseSuccessful = true;
+        _account = AppointmentsAccount.fromJson(JsonUtils.decodeMap(responseString));
+      } else {
+        Log.w('Failed to load Appointments Account. Response:\n$responseCode: $responseString');
+        _isLastAccountResponseSuccessful = false;
+        _account = null;
+      }
+    } else {
+      _isLastAccountResponseSuccessful = null;
+      _account = null;
+    }
+    NotificationService().notify(notifyAppointmentsAccountUpdated);
+  }
+
+  Future<void> _initAccountOnFirstSignIn() async {
+    // By Default
+    AppointmentsAccount defaultNewAccount = AppointmentsAccount(
+        notificationsAppointmentNew: true, notificationsAppointmentReminderMorning: true, notificationsAppointmentReminderNight: true);
+    Map<String, String> headers = {'Content-Type': 'application/json'};
+    String? accountJsonString = JsonUtils.encode(defaultNewAccount.toJson());
+    String? url = "${Config().appointmentsUrl}/services/account";
+    http.Response? response = await Network().post(url, auth: Auth2(), body: accountJsonString, headers: headers);
+    int? responseCode = response?.statusCode;
+    String? responseString = response?.body;
+    if (responseCode == 200) {
+      _isLastAccountResponseSuccessful = true;
+      _account = AppointmentsAccount.fromJson(JsonUtils.decodeMap(responseString));
+    } else {
+      Log.w('Failed to init default Appointments Account. Response:\n$responseCode: $responseString');
+      _isLastAccountResponseSuccessful = false;
+      _account = null;
+    }
+    NotificationService().notify(notifyAppointmentsAccountUpdated);
+  }
+
+  void changeAccountPreferences({bool? newAppointment, bool? morningReminder, bool? nightReminder}) {
+    if (_account != null) {
+      bool changed = false;
+      if (newAppointment != null) {
+        _account!.notificationsAppointmentNew = newAppointment;
+        changed = true;
+      }
+      if (morningReminder != null) {
+        _account!.notificationsAppointmentReminderMorning = morningReminder;
+        changed = true;
+      }
+      if (nightReminder != null) {
+        _account!.notificationsAppointmentReminderNight = nightReminder;
+        changed = true;
+      }
+      if (changed) {
+        _updateAccount();
+      }
+    }
+  }
+
+  Future<void> _updateAccount() async {
+    String? accountJsonString = JsonUtils.encode(_account);
+    Map<String, String> headers = {'Content-Type': 'application/json'};
+    String? url = "${Config().appointmentsUrl}/services/account";
+    http.Response? response = await Network().post(url, auth: Auth2(), body: accountJsonString, headers: headers);
+    int? responseCode = response?.statusCode;
+    String? responseString = response?.body;
+    if (responseCode == 200) {
+      _isLastAccountResponseSuccessful = true;
+      _account = AppointmentsAccount.fromJson(JsonUtils.decodeMap(responseString));
+    } else {
+      Log.w('Failed to init default Appointments Account. Response:\n$responseCode: $responseString');
+      _isLastAccountResponseSuccessful = false;
+      _account = null;
+    }
+    NotificationService().notify(notifyAppointmentsAccountUpdated);
+  }
+
   // ExploreJsonHandler
   @override bool exploreCanJson(Map<String, dynamic>? json) => Appointment.canJson(json);
   @override Explore? exploreFromJson(Map<String, dynamic>? json) => Appointment.fromJson(json);
@@ -189,6 +313,24 @@ class Appointments with Service implements ExploreJsonHandler, NotificationsList
   void onNotification(String name, dynamic param) {
     if (name == DeepLink.notifyUri) {
       _onDeepLinkUri(param);
+    } else if (name == AppLivecycle.notifyStateChanged) {
+      _onAppLivecycleStateChanged(param);
+    } else if (name == Auth2.notifyLoginChanged) {
+      _initAccount();
+    }
+  }
+
+  void _onAppLivecycleStateChanged(AppLifecycleState? state) {
+    if (state == AppLifecycleState.paused) {
+      _pausedDateTime = DateTime.now();
+    }
+    else if (state == AppLifecycleState.resumed) {
+      if (_pausedDateTime != null) {
+        Duration pausedDuration = DateTime.now().difference(_pausedDateTime!);
+        if (Config().refreshTimeout < pausedDuration.inSeconds) {
+          _loadAccount();
+        }
+      }
     }
   }
 }